+++ extension/app/src/components/conversation/usePublicConversation.ts
@@ -1,5 +1,6 @@
 import type { ConversationPublicType } from "@dust-tt/client";
-import { useAuthErrorCheck } from "@extension/lib/auth";
+import { useAuthErrorCheck } from "@extension/hooks/useAuthErrorCheck";
+import {} from "@extension/lib/auth";
 import { useDustAPI } from "@extension/lib/dust_api";
 import { useSWRWithDefaults } from "@extension/lib/swr";
 import type { KeyedMutator } from "swr";
